Attempting a big refactor
