+++ src/websocket.rs
@@ -1,2 +1,1 @@
 pub mod parser;
-pub mod client;
